# swagger.yaml
swagger: "2.0"
info:
  description: "Sqre server api"
  version: "1.0.0"
  title: "Sqre"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "yaroslav.ananchenko@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "sqre.ru"
basePath: "/"

schemes:
- "http"
- "https"

paths:

  /{link}/info:
    get:
      tags:
      - "Object"
      summary: "Get object by link"
      description: "Returns a single object"
      
      produces:
      - "application/json"
      parameters:
      - name: "link"
        in: "path"
        description: "Link code"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Info"
        "404":
          description: "Object not found"


  
definitions:

  Info:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      catid:
        type: "integer"
        description: "Category id"
        format: "int64"        
      typeid:
        type: "integer"
        description: "Type object id"
        format: "int64"        
      name:
        type: "string"
        example: "Some object name"
      anonce:
        type: "string"
        example: "Short about object"
      desc:
        type: "string"
        example: "Description object"
      url:
        type: "string"
        example: "https://ya.ru"       
        description: "Default url to redirect"
      lat:
        type: "number"
        format: "float"
        example: "60.3137422912492"
        description: "latitude"
      lon:
        type: "number"
        format: "float"
        example: "28.6288770629748"
        description: "longitude"
      version:
        type: "string"
        example: "0.1.45"       
        description: "Version of server software"        
      urls:
        type: "array"
        items:
          $ref: "#/definitions/SubArray"
      photos:
        type: "array"
        items:
          $ref: "#/definitions/SubArray"
      emails:
        type: "array"
        items:
          $ref: "#/definitions/SubArray" 
      maps:
        type: "array"
        items:
          $ref: "#/definitions/SubArray"           
      videos:
        type: "array"
        items:
          $ref: "#/definitions/SubArray"           
      audio:
        type: "array"
        items:
          $ref: "#/definitions/SubArray" 
      tracks:
        type: "array"
        items:
          $ref: "#/definitions/SubArray"           

  SubArray:
    type: "object"
    properties:
      name:
        type: "string"
      info:
        type: "string"



externalDocs:
  description: "Find out more about Sqre"
  url: "http://sqre.ru"